SARIMA: do not block waiting for result for live data
When we have to fit a SARIMA model (i.e. do a fresh batch of
predictions), this can take a number of seconds for each
series if the series is reasonably high frequency (e.g. bgp at
5 minute intervals).

Waiting for each preceding series to complete its model fits
before moving on to the next batch of live data can lead to our
kafka session timing out, which kills the RealTime source.

Instead, we're better off just handing off the live data and
letting libchocolatine's asynchronous processing handle it.
The result will be returned eventually, as long as we have
a continuous stream of live data to keep the generator looping.