AggSum: skip first timestamp by default, if it is partial
Now that we have a event detector that does not require a
learning phase, we need to be more careful about not passing
on partial aggregates that are not representative of a
typical measurement.

In the case of the first measurement, it may be that a previous
watchtower instance had read some of the probe results for that
timestamp so the first "sum" produced by this module would be
much lower than it should be. This can trigger an event if
using detectors like SArima, which prime themselves by reading
historical data prior to the first measurement and therefore
apply their detection algorithm immediately.