Update Historical source to work with inetintel API deployment
This is a bit hacky -- it only really work with a single
series and you need to include both entity type and code in the
url parameter, e.g.

http://api.ioda.inetintel.cc.gatech.edu/v2/signals/raw/country/DE

Then the other parameters can be provided via "queryparams", such
as datasource, maxPoints and sourceParams.

We mainly use this for debugging right now, but we may want to
do large backfills across many series in the future so this source
will need more attention then...