SArima: add a minimum series value threshold for alerting
This helps avoid generating alerts for series that are usually
very low values (e.g. < 5 ), where even a small drop can appear
to be a "high" magnitude event.

For SArima alerting, we probably just want to ignore those
series entirely as the alerts that we would get are not
significant in reality and will generally just obscure real
events behind their noise.