SARIMA: don't halt until all queued data has been processed
This is especially important when using sources that provide
finite amounts of historical data, as the data generator will
"end" when all of the data has been passed on to the detectors.

If we don't wait for the detectors to process that data then
we will exit without getting useful results for most of the
input data.