Fix excessive event generation resulting from latest changes
The hard-coded val of 0.5 was causing very minor anomalies
that were "detected" by SARIMA to be alerted by AlertKafka
as it did not realise that the intention was to simply
maintain its current "alert" state.

Now, we use the minimum observed val since SARIMA first
noticed the anomaly, so if the anomaly has always been minor
then the magnitude that reaches AlertKafka will continue to
be minor (and therefore not erroneously cause an alert to
be created).