Realtime source: ignore non-text kafka messages
If, for some reason, we happen to inject bogus messages into
the kafka topic that watchtower is trying to read from, we should
simply ignore those messages and try to carry on rather than
just crashing.

This change was motivated by the BGP edges consumer for GRIP
writing messages into the wrong topic and causing the BGP
watchtower to continually fail due to being unable to parse
the unexpected message.